requirements.txt
accelerate==1.13.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
ascii-colors==0.11.20
attrs==25.4.0
audioop-lts==0.2.2
av==16.1.0
beautifulsoup4==4.14.3
boto3==1.42.65
botocore==1.42.65
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorlog==6.10.1
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.9.0
cycler==0.12.1
dill==0.4.1
distro==1.9.0
doclayout-yolo==0.0.4
fast-langdetect==0.2.5
fastapi==0.135.1
fasttext-predict==0.9.2.4
ffmpy==1.0.0
filelock==3.25.1
flatbuffers==25.12.19
fonttools==4.62.0
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
google-api-core==2.30.0
google-auth==2.49.0
google-genai==1.66.0
googleapis-common-protos==1.73.0
gradio==5.49.1
gradio-client==1.13.3
gradio-pdf==0.0.22
groovy==0.1.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.6
httpx-sse==0.4.3
huggingface-hub==0.36.2
idna==3.11
imageio==2.37.3
jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
json-repair==0.58.5
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-loader==0.5
lightrag-hku==1.4.10
loguru==0.7.3
magika==1.0.2
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
mineru==2.7.6
mineru-vl-utils==0.1.22
modelscope==1.34.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.7.1
nano-vectordb==0.0.4.3
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
omegaconf==2.3.0
onnxruntime==1.24.3
openai==2.26.0
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pdfminer-six==20260107
pdftext==0.6.3
pillow==11.3.0
pipmaster==1.1.1
polars==1.38.1
polars-runtime-32==1.38.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1-modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.11.10
pydantic-core==2.33.2
pydantic-settings==2.13.1
pydub==0.25.1
pydyf==0.12.1
pygments==2.19.2
pyjwt==2.11.0
pyparsing==3.3.2
pypdf==6.8.0
pypdfium2==4.30.0
pyphen==0.17.2
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
pyyaml==6.0.3
qwen-vl-utils==0.0.14
raganything==1.2.9
referencing==0.37.0
regex==2026.2.28
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
robust-downloader==0.0.2
rpds-py==0.30.0
rsa==4.9.1
ruff==0.15.5
s3transfer==0.16.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.2.3
setuptools==82.0.1
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sse-starlette==3.3.2
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.1.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
ultralytics==8.4.21
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.23.0
zopfli==0.4.1
